sql queries from postgres to postgres_queries
