+++ app/modules/postgres.py
@@ -8,7 +8,6 @@ from time import time
 from psycopg2 import connect, sql, errors
 from dotenv import load_dotenv
 import logging
-import pandas as pd
 
 # load environments from .env file
 load_dotenv()
@@ -25,72 +24,13 @@ conn = connect(
 # initialize database cursor
 cursor = conn.cursor()
 
-def sql_from_dropdown(select: list[str], table: str, group_by: str, order_by: str, order: bool = False, limit: int = 10) -> list[tuple]:
-    """
-    This function gets the parameters provided by the dashboard, converts it to a sql query
-    and returns the result as a list of tuples.
-    Parameters:
-    - select: list[str] => the columns to be selected
-    - table: str => the table to select from
-    - group_by: str => the column to group by
-    - order_by: str => the column to order by
-    - order: bool = False => ASC or DESC order (True = ASC)
-    - limit: int = 10 => Row limitation
-    Return:
-    - list[tuples] => the result of the query as list of tuples
-    """
-
-    # check types
-    if not isinstance(select, (list, str)):
-        raise TypeError("select must be of type: list[str]")
-
-    if not isinstance(table, str):
-        raise TypeError("table must be of type: str")
-
-    if not isinstance(group_by, str):
-        raise TypeError("group_by must be of type: str")
-
-    if not isinstance(order_by, str):
-        raise TypeError("order_by must be of type: str")
-
-    if not isinstance(order, bool):
-        raise TypeError("order must be of type: bool")
-
-    if not isinstance(limit, int):
-        raise TypeError("limit must be of type: int")
-
-    sub_query = sql.SQL("""
-        SELECT {table_id}, COUNT({table_id}) as count 
-        FROM {entry_table} 
-        GROUP BY {table_id} 
-        ORDER BY {order_by} {order} 
-        LIMIT {limit}
-        """).format(
-        entry_table = sql.Identifier(f'entry_{table}'),
-        table_id = sql.Identifier(f'{table}_id'),
-        order_by = sql.SQL(order_by),
-        order = sql.SQL("ASC" if order else "DESC"),
-        limit = sql.Literal(str(limit))
-    )
-
-    # generate sql query
-    sql_query = sql.SQL("""
-        SELECT {select} 
-        FROM ({sub}) as sub_col 
-        LEFT JOIN {table} ON {on}
-        """).format(
-        select = sql.SQL(",").join(map(sql.SQL, select)),
-        sub = sub_query,
-        table = sql.Identifier(table),
-        on = sql.SQL(f'sub_col.{table}_id = {table}.id')
-    )
-
+def execute_query(query, name: str) -> list[tuple]:
     # execute query
     try:
         t1 = time()
-        cursor.execute(sql_query)
+        cursor.execute(query)
         t2 = time()
-        logging.debug(f'sql_from_dropdown query took: {t2 - t1:.3} seconds')
+        logging.debug(f'{name} query took: {t2 - t1:.3} seconds')
     except errors.SyntaxError as err:
         logging.warning(err)
     except Exception as err:
@@ -319,56 +259,3 @@ def paper_schools(key: str) -> list[tuple]:
 
     # return results
     return cursor.fetchall()
-
-def update_year_dropdown():
-    sql_query = """
-    SELECT DISTINCT EXTRACT(YEAR FROM mdate) as year
-    FROM entry
-    WHERE EXTRACT(YEAR FROM mdate) != EXTRACT(YEAR FROM NOW())
-    """
-    cursor.execute(sql_query)
-    years = cursor.fetchall()
-    options = [{"label": str(year[0]), "value": str(year[0])} for year in years]
-    return options
-
-def papers_per_month(year: str) -> pd.DataFrame:
-    sub_query = sql.SQL("""
-        SELECT entry_key, 'phdthesis' as entryType
-        FROM phdthesis
-        UNION
-        SELECT entry_key, 'mastersthesis' as entryType
-        FROM mastersthesis
-        UNION
-        SELECT entry_key, 'article' as entryType
-        FROM article
-        UNION
-        SELECT entry_key, 'book' as entryType
-        FROM book
-    """)
-
-    sql_query = sql.SQL("""
-        SELECT DATE_TRUNC('month', e.mdate) as month, n.entryType, COUNT(n.entryType)
-        FROM entry e
-        JOIN ({sub_query}) n ON e.key = n.entry_key
-        WHERE EXTRACT(YEAR FROM e.mdate) = {year}
-        GROUP BY month, n.entryType
-        ORDER BY month
-    """).format(
-        sub_query=sub_query,
-        year=sql.Literal(year)
-    )
-
-    try:
-        t1 = time()
-        cursor.execute(sql_query)
-        t2 = time()
-        logging.debug(f'papers_per_month query took: {t2 - t1:.3} seconds')
-    except errors.SyntaxError as err:
-        logging.warning(err)
-    except Exception as err:
-        logging.error(err)
-
-    results = cursor.fetchall()
-    df = pd.DataFrame(results, columns=['month', 'entryType', 'count'])
-
-    return df
